fix: insert data fix: events ui fix apiClientValidation feat: events dates

+++ projects/insertData/src/insertData.js
@@ -34,12 +34,16 @@ export class InsertData {
             payroll.link = data.link
             payroll.urlDownload = data.urlDownload
             payroll.income ??= 0
+            if (payroll.income === 'Honorífico') {
+                payroll.isHonorific = true
+                payroll.income = 0
+            }
             payroll.document = data.document ?? null
-            payroll.isDocumentValid = data.isDocumentValid ?? null
             payroll.confidences = JSON.stringify(data.confidences)
             if(payroll.position?.includes('regidor')) payroll.position = 'Regidor'
             return payroll
         })
+        if(payrolls.length <= 0) return
         await this.eventRepository.delete({...data,date: this.formatLastDayOfMonth(data.year, data.month)})
         await this.eventRepository.save(payrolls)
     }
